Rename and reorganize package
