+++ backend/src/main/java/com/cartowiki/webapp/authentication/model/request/LogInRequest.java
@@ -1,4 +1,4 @@
-package com.cartowiki.webapp.usermanagement.model.request;
+package com.cartowiki.webapp.authentication.model.request;
 
 /**
  * Data template for log in requests
